lruttl: potentially squash sema/eviction errors from logs
A couple of users have reported seeing:

  sema was closed but state is still pending

and/or

  lruttl: shaping_data did not find anything to evict, target was 1

messages in the diagnostic logs.

These arise due to the capacity of 1 that was configured
for the shaping data cache in the shaping helper and a race
condition where lookups are interleaved around a config epoch
bump triggered either by the initial epoch bump on startup,
or a bump produced by the TSA daemon.

This commit adjust the strategy for handling those errors if/when
they arise by attempting to restart the cache lookup process.

In addition, this commit raises the capacity of the shaping data
cache to avoid this point of contention.