rfc5321 client: fix openssl verify issue when hostname has trailing dot
I'm not totally sure why this isn't universally broken when using
openssl (instead of rustls), but in the specific case we were
investigating, the destination was configured via a routing_domain
and the resulting mx_host name had the trailing FQDN dot on it.

Removing that dot allows the certificate to verify, so let's
ensure that we strip it here in the client.