smtp_server: slim down EsmtpListenerParams
This commit removes the local cached copy of connection_gauge and
connection_denied_counter from EsmtpListenerParams.  There's not
a huge benefit from caching these in the listener params, and
I'd like to make the listener configuration more flexible, so even
if the cache was useful, this spot isn't the best spot for it.