smtp_server: introduce GenericEsmtpListenerParams type
This type is similar to the prior EsmtpListenerParams type,
except that every field is optional with no specified default.

This allows us to create an initial concrete set of parameters
(that hold the defaults), then progressively layer user specified
values from different context over the top of it.

In this commit we just have a single "base" context, which is
the set of parameters defined by start_esmtp_listener.

In the future, more contexts will be factored in.